apps/backend/requirements.txt
# Django
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-csp==3.8

# Database
psycopg[binary]==3.3.2
prisma==0.11.0

# Authentication
clerk-backend-api==5.0.0
PyJWT==2.9.0
bcrypt==5.0.0
cryptography==42.0.5
pyotp==2.9.0
qrcode[pil]==7.4.2

# Caching and Queue
redis==4.6.0
celery==5.3.6
django-ratelimit==4.1.0

# Email
resend==0.7.0

# AWS S3
boto3==1.34.34

# API Documentation
drf-spectacular==0.27.1

# Error Tracking and Monitoring
sentry-sdk[django,celery,redis]==1.40.0

# Application Performance Monitoring (APM)
# Install one of the following based on your APM provider:
# For New Relic: newrelic
# For DataDog: ddtrace datadog
newrelic==9.5.0  # New Relic APM agent
# ddtrace==2.5.0  # DataDog APM agent (uncomment if using DataDog)
# datadog==0.48.0  # DataDog metrics client (uncomment if using DataDog)

# Utilities
python-dotenv==1.0.1
bleach==6.1.0
requests==2.31.0
httpx==0.28.1

# Testing
pytest==7.4.4
pytest-django==4.7.0
pytest-asyncio==0.21.1
pytest-cov==4.1.0
hypothesis==6.98.3
nest-asyncio==1.6.0
